tests/wpt/tests/html/canvas/element/text/2d.text.drawing.style.absolute.spacing.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>Canvas test: 2d.text.drawing.style.absolute.spacing</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.text.drawing.style.absolute.spacing</h1>
<p class="desc">Testing letter spacing and word spacing with absolute length</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("Testing letter spacing and word spacing with absolute length");
_addTest(function(canvas, ctx) {

  _assertSame(ctx.letterSpacing, '0px', "ctx.letterSpacing", "'0px'");
  _assertSame(ctx.wordSpacing, '0px', "ctx.wordSpacing", "'0px'");

  ctx.letterSpacing = '3px';
  _assertSame(ctx.letterSpacing, '3px', "ctx.letterSpacing", "'3px'");
  _assertSame(ctx.wordSpacing, '0px', "ctx.wordSpacing", "'0px'");

  ctx.wordSpacing = '5px';
  _assertSame(ctx.letterSpacing, '3px', "ctx.letterSpacing", "'3px'");
  _assertSame(ctx.wordSpacing, '5px', "ctx.wordSpacing", "'5px'");

  ctx.letterSpacing = '-1px';
  ctx.wordSpacing = '-1px';
  _assertSame(ctx.letterSpacing, '-1px', "ctx.letterSpacing", "'-1px'");
  _assertSame(ctx.wordSpacing, '-1px', "ctx.wordSpacing", "'-1px'");

  ctx.letterSpacing = '1PX';
  ctx.wordSpacing = '10PX';
  _assertSame(ctx.letterSpacing, '1px', "ctx.letterSpacing", "'1px'");
  _assertSame(ctx.wordSpacing, '10px', "ctx.wordSpacing", "'10px'");

});
</script>
